accel-config: Add support for new attributes related to read buffers
This change is to match the driver change of attribute names. The bw
"token" attribute names are deprecated and are renamed as "read
buffers".

Add new libarary APIs to read and write the new attributes related to
read buffers. All internal implementation uses the "read buffer" name.
The external API will continue to support the "token" name and will be
deprecated. Writes to sysfs will attempt to read/write to "read buffer"
attributes and if not found in the case of older drivers, will fall back
to "token" attribute names.

Change internal implemenations to use read_buffer name instead of token.

Signed-off-by: Ramesh Thomas <ramesh.thomas@intel.com>
Reviewed-by: Dave Jiang <dave.jiang@intel.com>